Simplify service
The service provides a certain structure.  We imposed the structure to
get started and saw it as a value.  It turns out not all plugins
benefit from such structure and it may be easier to reason about a
plugin without it.  This is largely due to a much cleaned API space.

There is still room for improvement and we intend to tackle a bunch of
it before releasing this revamped generator.