Removal of the mystical who
The this.who thing was passed along and it seems people don't fully
grok what the meaning was.  The variable is removed now.  It has not
been replaced by a global constant like we used to do but rather by
inline strings.  With some luck these strings will be more obvious to
think about (and otherwise we change it again in the future).