Move normalizeLocation into a utility
The normalizeLocation function should not be duplicated in each
plugin.  For the time  being we're adding it to a utility so it's at
least not as much "in your face" as it used to be.  Migration will
thereby be minimal as we'll probably only need to change the import
path.